% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATclus_Ward.R
\name{MATclus_Ward}
\alias{MATclus_Ward}
\title{MATclus_Ward: Clustering jerárquico con método de Ward}
\usage{
MATclus_Ward(data, ..., k = 0, silueta = FALSE)
}
\arguments{
\item{data}{Dataframe con los datos.}

\item{...}{Variables específicas (sin comillas) a incluir en el clustering.}

\item{k}{Número de clusters a formar.}

\item{silueta}{Booleano, si TRUE, calcula el número óptimo de clusters.}
}
\value{
Lista con dendrograma, resumen de grupos y gráficos.
}
\description{
Realiza un análisis de clúster jerárquico usando el método de Ward y permite determinar el número óptimo de grupos.
}
