% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATcor.R
\name{MATcor}
\alias{MATcor}
\title{MATcor: Matriz de correlación gráfica}
\usage{
MATcor(data, ...)
}
\arguments{
\item{data}{Dataframe con los datos.}

\item{...}{Variables específicas (sin comillas) a incluir en la matriz de correlación.}
}
\value{
Un gráfico de correlación guardado en el entorno global.
}
\description{
Esta función crea una matriz de correlación gráfica entre las variables métricas de un dataframe.
Permite seleccionar un subconjunto específico de variables.
}
