% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATpca.R
\name{MATpca}
\alias{MATpca}
\title{MATpca: Análisis de Componentes Principales}
\usage{
MATpca(data, ...)
}
\arguments{
\item{data}{Dataframe con los datos.}

\item{...}{Variables específicas (sin comillas) a incluir en el análisis.}
}
\value{
Lista con tablas de resumen y gráficos.
}
\description{
Realiza un Análisis de Componentes Principales (PCA) sobre variables métricas.
Devuelve tablas de resumen y gráficos de autovalores y varianza acumulada.
}
