% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATpcascores.R
\name{MATpcascores}
\alias{MATpcascores}
\title{MATpcascores: Obtención de puntuaciones de componentes principales}
\usage{
MATpcascores(data, pca_object, Nscores = NULL, Rank = 10)
}
\arguments{
\item{data}{Dataframe de entrada (puede ser diferente al original).}

\item{pca_object}{Objeto resultante de prcomp().}

\item{Nscores}{Número de componentes principales a extraer (por defecto todas).}

\item{Rank}{Número de casos a mostrar en tablas y gráficos (por defecto 10).}
}
\value{
Lista con el dataframe de puntuaciones, tabla formateada y gráfico.
}
\description{
Calcula y devuelve puntuaciones de componentes principales de un PCA previamente realizado.
}
